Overhaul mount ID allocation: drop upper bound, add unique IDs
Replace the 10000-cap `IdAlloc` with `SparseIdAlloc` over
`[1, i32::MAX]`; exhaustion now surfaces as `ENOSPC`.

Add a monotonically increasing `unique_id: u64` alongside the
recyclable `id: u32` (mirrors Linux's `mnt_id_unique` and `mnt_id`),
looked up via the owning namespace's `mounts` table. `statx` returns
the unique ID when the caller sets `STATX_MNT_ID_UNIQUE`.

Bundle the two IDs in a `MountId` value: allocation is fallible,
and the recyclable half is returned to the pool when the owning
`Mount` is dropped.

+++ kernel/src/fs/fs_impls/procfs/pid/task/mountinfo.rs
@@ -44,10 +44,11 @@ pub(super) fn make_mount_point_path(
 
 /// A single entry in the mountinfo file.
 struct MountInfoEntry<'a> {
-    /// A unique ID for the mount (but not guaranteed to be unique across reboots).
-    mount_id: usize,
-    /// The ID of the parent mount (or self if it has no parent).
-    parent_id: usize,
+    /// The recyclable mount ID (matches [`Mount::id`]); reused after the
+    /// mount is dropped.
+    mount_id: u32,
+    /// The recyclable ID of the parent mount (or self if it has no parent).
+    parent_id: u32,
     /// The major device ID of the filesystem.
     major: u32,
     /// The minor device ID of the filesystem.
